refactor: Merge clustering canvas
