+++ src/visualization/__init__.py
@@ -1,3 +1,2 @@
 from .a_priori_canvas import APrioriGauge, APrioriGraphPlot, APrioriScatterPlot
-from .gmm_canvas import GMMCanvas
-from .k_means_canvas import KMeansCanvas
+from .clustering_canvas import ClusteringCanvas
